backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-jose[cryptography]
requests
python-multipart

